fix: rewrite for lossless artworks 0041 migration
This migration contained some unneeded code and did only work on empty
databases or as long as no artwork collections containing links to
artworks have been created. The current fix makes sure that the
migration works without errors even on a fully populated database and
that no data is lost in forward or backward migration.