refactor: simplify migration 0041
The initial migration before this fix was strangely convoluted and did
not work if there was existing data for the folders
(artworkcollectionmembership). Therefore it was reworked. But it turns
out that Django's basic operations suffice just as well, so this is the
migration Django initially should have created to rename the two models
and add add an artworks manytomany field.