fix: create slides content from relations
Once the slides field is introduced, we have to use the relationship
information from the AlbumMembership to generate and store the slides
representation of every Album. This fix adds the code to do that as a
separate operation in the same migration file where the slides field is
added to the Album model.