feat: add configurable django-cors-headers
Until now we did not need CORS headers for image, as the few AJAX
requests came directly from the pages served by Django. Now with the API
and a new frontend in the making, we need to set CORS headers
correspondingly. This is solved here in a similar was as in Showroom and
Portfolio, although only the most minimal set of needed directives is used.