build(deps): remove unused debug_toolbar
This was configured, but weirdly turned off for all cases with a custom
lambda function for SHOW_TOOLBAR_CALLBACK. While the __debug__ paths
still would be available the toolbar itself is never shown. And as we do
not facilitated the debug toolbar for API development, it is better to
remove it, until we have a better reason to add it again.