fix: gather featured artworks data directly from slide order
The logic how the featured artworks were assembled was a bit twisted.
Due to the __in filter and the following slice on the query set, the
order of the artworks did not correlate with the order of ids appearing
in the slides.

The update logic here only gathers up to num_artworks artworks, based on
the actual slide order (leaving out any duplicates).

Refs: #2778